refactor model helpers into conftest
